Moved ActivityExportingDelegate to its own package.
